Make the results of IvyResolve be target-dependent.
Previously the classpath file was always the latest one computed, even
if you then start working on a different target whose BUILD files
haven't change (and therefore won't trigger a new resolve).

This change qualifies each classpath file with (a hash of the ids of)
the targets it was generated from, and the main classpath file is now
a symlink to the appropriate one.

Signed-off-by: John Sirois <jsirois@twitter.com>