Zinc integration into pants.
Zinc is the new standalone incremental scala compiler, refactored
out of SBT.

The most tricky part of this change is the interplay between incremental
compilation and non-flat compilation. Each compiler pass has an analysis
cache associated with it. If building non-flat, each pass has to receive
its upstream dependencies as separate dependencies on the classpath,
that are then associated with the appropriate analysis caches.

This means that in this case each pass must emit .class files to a separate
directory. We then copy those over to the main classes directory.