Generate error for classes without package.
The implementation of zinc rebase/split/merge requires that all scala
source files have declared packages. In the current code, when a user
builds a scala source file without a package declaration, pants crashes
and dumps its stack with a "No such file or directory" error.

This change adds code to detect when a compilation produced a class
file without a package, and generates a comprehensible error message.

Auditors: benjy