[PATCH 070/153] Remove swallowed exceptions in git branch context manager
Currently there is no way to raise an error inside a git branch
context manager that will propagate outside of the branch. This is a
problem for the build module of the mesos client: a task create/update
happens regardless of if the build throws.

RB_ID=178281

(sapling split of a972b504f66ec05ea25d459aa4381d877aec302e)