[PATCH 3467/7508] metrics: Introduce HistogramInterface and Snapshot
Also reverse the model of Histogram, previously creating an Histogram would register Gauges in the main registry, now it doesn't anymore, but the main registry is taking a snapshot of the Histogram. It has the advantages of
solving the problem of inconsistant percentiles (in the past we can query one percentile, add a bunch of values, and query another percentile).

RB_ID=238509

(sapling split of 41e4bb9744d5644b08a464f8862077ae37920ba2)