Rename BootstrapUtils to something more dignified.
It's a proper class that models some entity, "utils" is not
a good name for it.

Also renamed its methods to be a bit shorter and more readable.

This is all because my upcoming refactoring of {java|scala}_compile.py
will be interacting quite significantly with this class, and
I want the names to make more sense.

Patrick: I was going to ask you about this first, but you
were in transit all day, afaict. Feel free to push back
if you're opposed to this change.

Auditors: pl