[PATCH 7497/7508] util: Introduce LowResClock, a fast but low resolution Clock.
- Problem

Some piece of code need frequent access to system time but doesn't really
require a high precision. The `Windowed` class is an example of that, we need
~second resolution but read time at every access.

- Solution

Introduce a LowResClock which wrap the system clock and cache its value and
start a periodic task responsible for updating its value. This replace the
syscall by reading a volatile variable.

- Result

The performance of system relying on `Windowed` is greatly improved.
As an example, the StatsInsertionBenchmark is more than two times
faster (not part of this review):
MetricsStatsInsertion   153.0 =
MetricsStatsInsertion   71.5 =

RB_ID=349819

(sapling split of 142c18407ff826ed5552d68198da5bba542a3492)