migrate-history.sh: Add -o subdir option
This option allows the usre to specify a subdirectory of the incoming
repository to be imported, instead of importing the whole thing.
Only commits which affect that subdirectory will be migrated.

Testing Done:
The test script passes (added a test for the new option)

Reviewed at https://rbcommons.com/s/twitter/r/715/