Improve commons CI.
Previously we were building under python2.6 and 2.7 even
though Twitter has since moved on to drop support for python2.6.

This change instead fixes the build to python2.7 and leverages
that pinning to pivot the matrix axis to jvms and exercise 1.6 & 1.7.

This change also removes the custom notification and falls back to the
default of emailing the submitter and comitter.  The idea here is
to give responsible contributors the feedback they need to roll back
or fix breaks instead of bottlenecking on 2 potentially unrelated
people.