Ensure liveness of reads in LowResClockTest.
Previously a bare FakeClock was used to model time in the test thread
and the test updater thread.  This arrangement did not account for
liveness of reads of the underlying FakeClock current time and as
evidenced by failing test runs, the read time could lag:

  1) testLowResClock(com.twitter.common.util.LowResClockTest)
  java.lang.AssertionError: expected:<1000> but was:<0>
    ...
    at com.twitter.common.util.LowResClockTest.testLowResClock(LowResClockTest.java:151)

This fix just wraps all time access with synchronized forwarders to
ensure liveness.

Also collapse the Token abstraction into direct use of CountDownLatches
for coordination signaling.