Upgrade to pantsbuild.pants 0.0.32.
This release brings in fixes for the idea and eclipse goals which were
previously broken with missing template resources.

It pulls in a similar fix for python-eval which is now enabled for the
repo and included in CI.  This provides a limited form of compile-time
checking that python target basics are in order:
+ checks that transitive BUILD deps are sufficient
+ validates that all top-level symbols are valid including function,
  method and class declarations
+ and decorators run at the their top level of performing function
  wrapping

These compile checks smoked out 3 targets with missing deps that are now
fixed.  It also identified 6 other special targets that do deliberate
dependency gymnastics of one form or another and these are excluded from
the "compile" as a result.

Finally a small config migration, simplification of the pants script and
killing of a few un-used build-support/python scripts round out the
cleanups.