Avoid ivy cache thrash due to ivydata updates.
The `ivydata-*.properties` and root-level `*.{properties,xml}` files'
effect on resolution time is in the noise, but they are re-timestamped
in internal comments and fields by ivy on each run and this leads to
cache thrash in travis-ci. Kill these files before the cache check to
avoid un-needed cache re-packing and re-upload which is a ~90s
operation that occurs at the end of our CI run.