Make args-apt resource mapping paths more robust.
Previously the code used FileObject.getName() which has no guarantees to
be anything but a friendly name.  Switch to using the FileObject URI
which has tighter semantics.  This fixes m

Previously the code used FileObject.getName() which has no guarantees to
be anything but a friendly name.  Switch to using the FileObject URI
which has tighter semantics.  This fixes mappings under Sun JDK
1.6.0_45-b06 for example and works with all four of OpenJDK & Oracle JDK
1.7.0_79 & 1.8.0_45.