Upgrade to pants 0.0.44.
The release notes are here:
  https://pypi.python.org/pypi/pantsbuild.pants/0.0.44

Of particular interest to twitter/commons is the graduation of the
isolated jvm compile strategy to a viable out-of-the-box solution for
java & scala compiles.  This change includes a switch over to that
strategy which necessitated an upgrade to scala 2.10.

Testing Done:
CI went green here:
  https://travis-ci.org/twitter/commons/builds/77066466

Bugs closed: 410

Reviewed at https://rbcommons.com/s/twitter/r/2691/