python: Add IP addresses to serverset entries
ServerSet entries contain hostnames that require a DNS lookup before
use. When hostnames map to a unique, fixed IP address, this additional
lookup is unnecessary and can put strain on DNS infrastructure in large
deployments.

This change adds 2 optional fields to server set endpoints: `inet` and
`inet6` respectively for a human readable representation of an IPv4 and
of an IPv6 address.

Hostnames are still mandatory, for backward compatibility with clients
that expect those fields. Port is untouched.